+++ database/migrations/2024_09_09_155225_create_experiences_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('experiences', function (Blueprint $table) {
             $table->id();
             $table->integer('lvl');
-            $table->integer('exp');
-            $table->integer('exp_diff');
+            $table->bigInteger('exp');
+            $table->bigInteger('exp_diff');
             $table->timestamps();
         });
     }
